Revert "fix: make MonotonicClock monotonic (#31)" (#42)
* fix: Undo the breaking toNanoseconds change in SwiftExtensions by @bklimt

* ci: add backwards compatibility check

* fix: run opentelemetry-swift tests

* revert 2.4.1 patch to prove backwards incompatibility check catches the breaking change

This reverts commit 2e6e5a26287f7bbe34018edaf5734ab81de5505e.

* restore linux test

* allow retry

* fix: Undo the breaking toNanoseconds change in SwiftExtensions by @bklimt

* invalidate cache

* Revert "fix: Undo the breaking toNanoseconds change in SwiftExtensions by @bklimt"

This reverts commit 1c47ce57f1314fc0f64a90cb1d08ea7babcb46f2.

* Revert "fix: make MonotonicClock monotonic (#31)"

This reverts commit 07c725fa66d3d946b037a6fbfe40f6541b6c5a6d.

* remove caching

* fix: make MonotonicClock monotonic (#31)

* fix: roll forward

* doc: add warnings to global extensions

---------

+++ Tests/OpenTelemetrySdkTests/Internal/MonotonicClockTests.swift
@@ -7,7 +7,7 @@ import OpenTelemetrySdk
 import XCTest
 
 class MonotonicClockTests: XCTestCase {
-  let epochNanos: Int64 = 1_234_000_005_678
+  let epochNanos: UInt64 = 1_234_000_005_678
   var testClock: TestClock!
 
   override func setUp() {
@@ -15,7 +15,7 @@ class MonotonicClockTests: XCTestCase {
   }
 
   func testNanoTime() {
-    XCTAssertEqual(testClock.now, TestUtils.dateFromNanos(epochNanos))
+    XCTAssertEqual(testClock.now, TestUtils.dateFromNanos(Int64(epochNanos)))
     XCTAssertEqual(testClock.nanoTime, epochNanos)
     let monotonicClock = MonotonicClock(clock: testClock)
     testClock.advanceNanos(12345)
